Update html example in first-steps.md
This change was needed when taking a vanilla node.js with express project template to get form posting to work properly out-of-box. As I an not that experience front-end developer, it took a while to find out why I wasn't getting the editor's content into the body of submitted form. Finally the only issue was to have also name in addition to id defined for the textarea, and after that everything worked as a breeze.