Fix minor inaccuracies in RTC snapshot handling
The snapshot callback is not executed at regular intervals, but instead
after edits are made to the document.

Snapshots are also stored on the server, but only as a editor checkpoint
so operations do not need to be replayed from the beginning of the
document editing session. By replacing "regular" with "HTML", the
statement is more correct. There may be some HTML snapshots currently
being sent to the server, but as legacy behaviour that is likely to be
removed. (In all cases, these snapshots are encrypted before being sent
to the server, and not readable without the client.)