6.0-release-notes-apis.adoc
`editor.fire()` section: added clause specifically noting the API has been re-named.

Also: apologies to @tiny-ben-tran. This was your edit/suggestion and, when I attempted to push it as such, the PR page returned an ‘outdated’ error, refreshed, and promptly hid your review so well I’ve not been able to find it on the page.