6.0-release-notes-apis.adoc and migration-from-5x.adoc
`fire` sections in both.

Expanded notes in both documents on `fire` being both re-named and marked as deprecated. Re-written as follows:

`fire` has been aliased to `dispatch`, so existing configurations should still work as expected.

`fire` has also, however, been marked as _deprecated_. Editing configurations to use `dispatch()` is recommended.

A useful reminder that brevity is not a virtue in and of itself.