modules/ROOT/partials/release-notes/6.0-release-notes-ui-and-ux-elements-and-components.adoc
Section on editor text entry height repeated to no useful effect.

Information on this change was extant in both the New and Improved and Changed sections.

The wording was slightly different, but the information was the same. The ‘Changed’ material was moved to the ‘New and Improved’ section and the original ‘Changed’ material was removed.