DOC-1821: mergetags.adoc, 6.2-release-notes.adoc, mergetags-menu-items.adoc, & mergetags_list.adoc
1. mergetags.adoc: replaced instances of ‘Merge Tags’ (however capitalised) with ‘{pluginname}’.

Attributes are defined and declared; I should use them.

Some instances were copy-edited to ‘merge tags’. The plugin is a Proper Noun and gets both words capitalised. But the particular strings that are placeholders for other data are specific merge tags, which are plain old noun phrases.

2. Corrected capitalisation instances through other files, most particularly, `6.2-release-notes.adoc`.

The same ‘if it is a placeholder string, it is a plain old noun phrase otherwise it is set as a Proper Noun’ treatment applied through these files as well.