DOC-1836: fullscreen.adoc & fullscreen_native.adoc
Added a NOTE admonition to `fullscreen.adoc` to keep this in sync with the equivalent admonition added to the 5.x docs consequent to the release of 5.10.6.

Plus a small copy-edit tweak to one of the bullet list items in the ‘Limitations of the fullscreen_native option’ section of `fullscreeen_native.adoc`.

Said copy-edit was noted when trawling through looking for the right spot to add the NOTE above.