DOC-1758: release-notes5106.md, nav.yml & events.md (#2542)
* DOC-1758: release-notes5106.md & fullscreen.md

Added complete draft of the 5.10.6 release notes as new file `release-notes5106.md`.

Also added NOTE admonition to `fullscreen.md` regarding change to this plugin introduced with the 5.10.6 update.

* Update release-notes/release-notes5106.md

typo correction: s/.././

Co-authored-by: Lee Newson <lee.newson@tiny.cloud>

* DOC-1758: release-notes5106.md & nav.yml

Edits as per reviews and added release-notes5106.md to nav.yml.

* DOC-1758: release-notes5106.md & nav.yml

Added new section — Accompanying Premium self-hosted server-side component changes — to `releasenotes5106.md` and added ToC entry to this section to `nav.yml`

Also added in-document link to this new section to the Overview section of `release-notes5106.md`.

* DOC-1758: fullscreen.md & events.md

Moved documentation of `ResizeEditor` event firing when `fullscreen` plugin toggles the fullscreen mode.

It was a NOTE admonition in`fullscreen.md`.

It is now part of the description field for the `ResizeEditor` event in the *Editor core events* table of `events.md`. (And is no longer an admonition.)

* DOC-1758: release-notes5106.md

Nomenclature edit: s/an object/editor content/.

* DOC-1758: fullscreen.md

A blank line was accidentally removed when undoing an edit. The blank line has been restored.

* Update release-notes/release-notes5106.md

Left-over from copy-and-paste edit typo correction.

Co-authored-by: Lee Newson <lee.newson@tiny.cloud>

* DOC-1758: events.md

Typo correction: s/toogles/toggles/.

* DOC-1758: events.md

Edited ResizeEditor event description to keep it in sync with the equivalent edit to events.adoc in the 6.x docs suite. Edit is a minor re-write of an edit suggested by @MitchCTiny.

* Update advanced/events.md

Missing pronoun — when — no longer missing.

Co-authored-by: Lee Newson <lee.newson@tiny.cloud>

* Update release-notes/release-notes5106.md

Co-authored-by: ltrouton <46954949+ltrouton@users.noreply.github.com>

* DOC-1758: release-notes5106.md

Corrected version number for Image proxy: s/2.106.1/2.106.4/.

Co-authored-by: Lee Newson <lee.newson@tiny.cloud>
Co-authored-by: ltrouton <46954949+ltrouton@users.noreply.github.com>