DOC-1777: 6.3-release-notes.adoc
Removed documentation of French Advanced Typography rules being a known issue in the Advanced Typography 1.0.0 plugin.

As per TINY-9438, French has been disabled in the initial release of the plugin, specifically because of this known issue (raised in TINY-9434).

Also deleted a commented-out Jira URL as no longer required in the section it was deleted from.